pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "whalesound-clusters"
version = "0.1.0"
description = "Exploratory bioacoustics: clustering humpback whale song units from SanctSound audio"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "whalesound_cluster", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.24.0"
scipy = "^1.11.0"
soundfile = "^0.12.1"
librosa = "^0.10.1"
pandas = "^2.0.0"
pyarrow = "^18.0.0"
umap-learn = "^0.5.4"
hdbscan = "^0.8.33"
scikit-learn = "^1.3.0"
streamlit = "^1.28.0"
plotly = "^5.17.0"
pyyaml = "^6.0.1"
google-cloud-storage = "^2.14.0"
tqdm = "^4.66.0"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
ruff = "^0.1.6"
mypy = "^1.7.0"
types-pyyaml = "^6.0.1"
types-click = "*"

[tool.poetry.scripts]
whale-download = "whalesound_cluster.cli:download_cmd"
whale-segment = "whalesound_cluster.cli:segment_cmd"
whale-featurize = "whalesound_cluster.cli:featurize_cmd"
whale-cluster = "whalesound_cluster.cli:cluster_cmd"
whale-ui = "whalesound_cluster.cli:ui_cmd"
whale-pipeline = "whalesound_cluster.cli:pipeline_cmd"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "DTZ", "T10", "EM", "ISC", "ICN", "G", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SIM", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["E501", "PLR0913", "PLR2004"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**" = ["ARG", "PLR2004"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/whalesound_cluster --cov-report=term-missing"

